refactor advisory policy pack detail router
