+++ src/app/routers/advisory_policy.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter, Path
+from fastapi import APIRouter
 
 from app.contracts.advisory_policy import AdvisoryPolicyEnvelopeResponse
 from app.middleware.correlation import correlation_id_var
@@ -18,33 +18,5 @@ router = APIRouter(prefix="/api/v1", tags=["advisory-policy"])
 )
 async def list_policy_packs() -> AdvisoryPolicyEnvelopeResponse:
     return await advisory_policy_service().list_policy_packs(
-        correlation_id=correlation_id_var.get(),
-    )
-
-
-@router.get(
-    "/advisory-policy-packs/{policy_pack_id}/versions/{policy_version}",
-    response_model=AdvisoryPolicyEnvelopeResponse,
-    summary="Get Advisory Policy Pack Version",
-    description=(
-        "Returns a source-owned policy pack version from lotus-advise, including its "
-        "supportability and activation posture when supplied by Advise."
-    ),
-)
-async def get_policy_pack_version(
-    policy_pack_id: str = Path(
-        ...,
-        description="Policy pack identifier owned by lotus-advise.",
-        examples=["policy_pack_sg_private_banking"],
-    ),
-    policy_version: str = Path(
-        ...,
-        description="Policy pack version identifier owned by lotus-advise.",
-        examples=["2026.05"],
-    ),
-) -> AdvisoryPolicyEnvelopeResponse:
-    return await advisory_policy_service().get_policy_pack_version(
-        policy_pack_id=policy_pack_id,
-        policy_version=policy_version,
         correlation_id=correlation_id_var.get(),
     )
